Deleted base classes and embedded their code in the corresponding classes.
