--- app/src/main/java/ar/com/wolox/android/bootstrap/ui/base/BaseView.kt
@@ -1,6 +1,0 @@
-package ar.com.wolox.android.bootstrap.ui.base
-
-interface BaseView {
-    fun showLoading()
-    fun hideLoading()
-}
