+++ app/src/main/java/ar/com/wolox/android/bootstrap/ui/root/RootViewModel.kt
@@ -1,8 +1,7 @@
 package ar.com.wolox.android.bootstrap.ui.root
 
+import androidx.lifecycle.ViewModel
 import ar.com.wolox.android.bootstrap.Constants
-import ar.com.wolox.android.bootstrap.ui.base.BaseView
-import ar.com.wolox.android.bootstrap.ui.base.BaseViewModel
 import ar.com.wolox.android.bootstrap.utils.SharedPreferencesManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
@@ -10,7 +9,7 @@ import javax.inject.Inject
 @HiltViewModel
 class RootViewModel @Inject constructor(
     private val sharedPreferencesManager: SharedPreferencesManager
-) : BaseViewModel<BaseView>() {
+) : ViewModel() {
 
     val isUserLogged: Boolean
         get() = sharedPreferencesManager[Constants.USER_IS_LOGGED_KEY, false]
